Rename internal lib pkg -> pkg per conventions.
