+++ pkg/config/test/revisions/test_array.yaml
