+++ pkg/core/sequence_test.go
