+++ pkg/core/test/duplicate_revisions/revisions/test_array.yaml
