+++ pkg/core/test/parse_test_dir/config.yml
