+++ pkg/core/test/parse_test_dir/revisions/test.yaml
