+++ pkg/core/test/parse_test_dir_yaml_revisions/revisions/test_array.yaml
