+++ pkg/core/testDirs.go
