+++ pkg/core/walk.go
