+++ pkg/core/walk_test.go
