+++ pkg/drivers/sqlite/sqlite.go
@@ -4,7 +4,7 @@ import (
 	"database/sql"
 	"encoding/json"
 
-	"github.com/PacketFire/immigrant/lib/core"
+	"github.com/PacketFire/immigrant/pkg/core"
 	_ "github.com/mattn/go-sqlite3"
 )
